Change all assetIds to be ulong

+++ dotnet-algorand-sdk/V2/Algod/Model/Model.cs
@@ -149,7 +149,7 @@ namespace Algorand.V2.Algod.Model
     {
         /// <summary>unique asset identifier</summary>
         [Newtonsoft.Json.JsonProperty("index", Required = Newtonsoft.Json.Required.Always)]
-        public BigInteger Index { get; set; }
+        public ulong Index { get; set; }
 
         [Newtonsoft.Json.JsonProperty("params", Required = Newtonsoft.Json.Required.Always)]
         [System.ComponentModel.DataAnnotations.Required]
@@ -171,7 +171,7 @@ namespace Algorand.V2.Algod.Model
 
         /// <summary>Asset ID of the holding.</summary>
         [Newtonsoft.Json.JsonProperty("asset-id", Required = Newtonsoft.Json.Required.Always)]
-        public BigInteger AssetId { get; set; }
+        public ulong AssetId { get; set; }
 
         /// <summary>\[f\] whether or not the holding is frozen.</summary>
         [Newtonsoft.Json.JsonProperty("is-frozen", Required = Newtonsoft.Json.Required.Always)]
